Add new "mode" config option
Add a new config option called "mode" inside the rewards configuration file. When set to "random", the player will be given a number of random rewards from the list specified by the "amount" option. When set to "all", the player will be given one of every reward in the list.